restructuring file structure
