+++ src/index.js
@@ -1,6 +1,6 @@
 const { ApolloServer } = require('apollo-server')
-const typeDefs = require('./schema')
-const resolvers = require('./resolvers')
+const typeDefs = require('./schema/schema')
+const resolvers = require('./schema/resolvers')
 
 require('dotenv').config()
